Feat : 카페 배경화면, 유저 프로필 사진 관련 기능 추가
카페 배경화면 추가, 수정 기능 추가
유저 프로필 사진 수정 기능 추가

+++ src/controllers/cafeController.js
@@ -12,16 +12,22 @@ export const getRegister = (req, res) => {
 };
 
 export const postRegister = async (req, res) => {
+  // form의 입력값 받아오기
   const { name, description, location, theme, level, rating } = req.body;
-  if (level <= 0 || level > 5) {
-    return res
-      .status(400)
-      .render(REGISTER_PAGE, { levelErrorMsg: '카페의 난이도는 1에서 5 사이입니다.' });
+  const { file } = req;
+  const backgroundUrl = file ? file.path : '';
+
+  // 카페의 난이도가 1~5 사이인지 확인
+  if ((level && level <= 0) || level > 5) {
+    return res.status(400).render(REGISTER_PAGE, {
+      levelErrorMsg: '카페의 난이도는 1에서 5 사이입니다.',
+    });
   }
+  // 카페의 평점이 1~10 사이인지 확인
   if ((rating && rating <= 0) || rating > 10) {
-    return res
-      .status(400)
-      .render(REGISTER_PAGE, { ratingErrorMsg: '카페의 평점은 1에서 10 사이입니다.' });
+    return res.status(400).render(REGISTER_PAGE, {
+      ratingErrorMsg: '카페의 평점은 1에서 10 사이입니다.',
+    });
   }
   const registeredCafe = await Cafe.create({
     name,
@@ -32,6 +38,7 @@ export const postRegister = async (req, res) => {
       level,
       rating,
     },
+    backgroundUrl,
   });
   res.redirect('/');
 };
@@ -55,17 +62,25 @@ export const getEdit = async (req, res) => {
 };
 
 export const postEdit = async (req, res) => {
+  // form의 입력값 받아오기
   const { cafeId } = req.params;
   const { name, description, location, theme, level, rating } = req.body;
+  const { file } = req;
+  let { backgroundUrl } = await Cafe.findById(cafeId);
+  backgroundUrl = file ? file.path : backgroundUrl ? backgroundUrl : '';
+
+  // 카페의 난이도가 1~5 사이인지 확인
   if (level < 1 || level > 5) {
-    return res
-      .status(400)
-      .render('cafe/edit', { levelErrorMsg: '카페의 난이도는 1에서 5 사이입니다.' });
+    return res.status(400).render('cafe/edit', {
+      levelErrorMsg: '카페의 난이도는 1에서 5 사이입니다.',
+    });
   }
+
+  // 카페의 평점이 1~10 사이인지 확인
   if (rating < 1 || rating > 10) {
-    return res
-      .status(400)
-      .render('cafe/edit', { ratingErrorMsg: '카페의 평점은 1에서 10 사이입니다.' });
+    return res.status(400).render('cafe/edit', {
+      ratingErrorMsg: '카페의 평점은 1에서 10 사이입니다.',
+    });
   }
   await Cafe.findByIdAndUpdate(cafeId, {
     name,
@@ -76,12 +91,17 @@ export const postEdit = async (req, res) => {
       level,
       rating,
     },
+    backgroundUrl,
   });
   res.redirect(`/cafes/${cafeId}`);
 };
 
 export const deleteCafe = async (req, res) => {
   const { cafeId } = req.params;
+  const { cafe } = await Cafe.findById(cafeId);
+  if (!cafe) {
+    return res.status(404).render('404');
+  }
   await Cafe.findByIdAndDelete(cafeId);
   res.redirect('/');
 };
